split out what is v6 in perpetuity vs what is latest
this is largely Luiz' suggestions in review (or me answering questions
by making code more obvious. hopefully.)
