+++ bin/propolis-server/src/lib/spec/api_spec_v2.rs
@@ -12,24 +12,21 @@ use propolis_api_types_versions::{
     v1, v1::instance::ReplacementComponent, v2, v3,
 };
 
-use super::{builder::SpecBuilder, Spec};
+use super::{builder::SpecBuilder, LegacyApiSpecError, Spec};
 use crate::migrate::MigrateError;
 
 #[cfg(feature = "failure-injection")]
 use super::MigrationFailure;
 
-// v2 does not introduce new opportunities for Spec->InstanceSpec conversion
-// to fail, so we can reuse the v1 error type directly.
-use super::api_spec_v1::ApiSpecError;
-
 impl TryFrom<Spec> for v2::instance_spec::InstanceSpec {
-    type Error = ApiSpecError;
+    type Error = LegacyApiSpecError;
 
     fn try_from(mut val: Spec) -> Result<Self, Self::Error> {
-        // A V2 InstanceSpec is just a V1 InstanceSpec with an optional `smbios_type1_input`.
-        // Emptying out the SMBIOS Type 1 input means this either can be converted to a V1 spec
-        // which we can losslessly make V2 by adding the SMBIOS table input back in, or we wouldn't
-        // be able to get to a V2 InstanceSpec either way.
+        // A V2 InstanceSpec is just a V1 InstanceSpec with an optional
+        // `smbios_type1_input`.  Emptying out the SMBIOS Type 1 input means
+        // this either can be converted to a V1 spec which we can losslessly
+        // make V2 by adding the SMBIOS table input back in, or we wouldn't be
+        // able to get to a V2 InstanceSpec either way.
         let smbios = val.smbios_type1_input.take();
 
         let v1::instance_spec::InstanceSpec { board, components } =
@@ -40,7 +37,7 @@ impl TryFrom<Spec> for v2::instance_spec::InstanceSpec {
 }
 
 impl TryFrom<v2::instance_spec::InstanceSpec> for Spec {
-    type Error = ApiSpecError;
+    type Error = LegacyApiSpecError;
 
     fn try_from(
         value: v2::instance_spec::InstanceSpec,
@@ -54,7 +51,7 @@ impl TryFrom<v2::instance_spec::InstanceSpec> for Spec {
 /// additional (non-v2) components to the builder before calling `finish()`.
 pub(crate) fn v2_to_spec_builder(
     value: v2::instance_spec::InstanceSpec,
-) -> Result<SpecBuilder, ApiSpecError> {
+) -> Result<SpecBuilder, LegacyApiSpecError> {
     let v3_spec: v3::instance_spec::InstanceSpec = value.into();
 
     crate::spec::api_spec_v3::v3_to_spec_builder(v3_spec)
